Add a build script for Python 2.7.12 with UCS-4 Unicode support
The default Python build's unicode mode (until Python 3.3, when PEP 393
landed) is UCS-2. However Ubuntu's system Python 2.7 is compiled in
UCS-4 mode. This new build script overrides the default for parity with
system Python, and to avoid hard to debug unicode bugs.

Ideally when Python 2.7.13 is released, the buildpack will use UCS-4
mode by default, however for now this is being kept as a separate
runtime version suffix, since:
* It seems unfair to switch the runtime under people, if they have
pinned to 2.7.12 expecting it to remain unchanged.
* Switching UCS modes invalidates existing Python package installs, so
we'd also have to purge site-packages in the cache manually, given that
the cache is only auto-purged if the runtime version in runtime.txt has
changed.

Fixes #305.